fix: event 행사 수정 이미지 교체 로직 오류 수정

+++ event/src/main/java/life/eventory/event/service/impl/EventServiceImpl.java
@@ -95,7 +95,8 @@ public class EventServiceImpl implements EventService {
         // 이미지 ID 설정
         if (newImageId != null) {
             event.setPosterId(newImageId);
-            communicationService.deletePoster(oldImageId);
+            if (oldImageId != null)
+                communicationService.deletePoster(oldImageId);
         }
 
         try {
